protected route for Theme.delete()

+++ resources/theme.py
@@ -3,6 +3,7 @@ from flask import request
 from models.theme import Theme
 from models.db import db
 from sqlalchemy.orm import joinedload
+from middleware import strip_token, read_token
 
 
 class Themes(Resource):
@@ -23,11 +24,18 @@ class OneTheme(Resource):
         theme = Theme.find_by_id(id)
         return theme.json()
 
-    def delete(self, id):
-        theme = Theme.find_by_id(id)
-        db.session.delete(theme)
-        db.session.commit()
-        return{"msg": 'Theme deleted', 'payload': theme.id}
+    def delete(sefl, id):
+        token = strip_token(request)
+        if token:
+            payload = read_token(token)
+            if payload != "Signature Invalid" and payload != "Invalid Token":
+                # check that id of authenticated user matches theme's user_id
+                theme = Theme.find_by_id(id)
+                if theme.user_id == payload['id']:
+                    db.session.delete(theme)
+                    db.session.commit()
+                    return {"msg": 'Theme deleted', 'payload': theme.id}
+        return 'Unauthorized', 401
 
     def put(self, id):
         theme = Theme.find_by_id(id)
